+++ src/consts.rs
@@ -2,4 +2,4 @@
 pub const DEFAULT_HYPR_CONFIG_PATH: &str = "~/.config/hypr/";
 
 /// The path theme repos are downloaded to
-pub const DEFAULT_DOWNLOAD_PATH: &str = "~/.local/share/hyprtheme/";
+pub const DEFAULT_DOWNLOAD_PATH: &str = "~/.config/hyprtheme/";
