--- src/modules.rs
@@ -1,1 +1,0 @@
-pub mod theme;
